Combined procedural scene menus
* Line stack and chain menu in one procedural scene menu with a combo
